requirements.txt
pandas==1.2.5
pdf2docx==0.5.1
Pillow==9.0.0
PyPDF2==1.26.0
